Feat: Add new date-picker component
