+++ src/components/date-picker/date-picker.styles.ts
@@ -1,0 +1,17 @@
+import { StyleSheet } from 'react-native';
+
+export const styles = StyleSheet.create({
+  container: {
+    marginVertical: 8,
+  },
+  button: {
+    paddingVertical: 12,
+    paddingHorizontal: 16,
+    backgroundColor: '#e0e0e0',
+    borderRadius: 6,
+    alignItems: 'center',
+  },
+  text: {
+    fontSize: 16,
+  },
+});
