+++ src/components/date-picker/date-picker.types.ts
@@ -1,0 +1,5 @@
+export interface DatePickerProps {
+  date: Date;
+  onDateChange: (newDate: Date) => void;
+  mode?: 'date' | 'time' | 'datetime';
+}
