+++ src/components/date-picker/index.ts
@@ -1,0 +1,2 @@
+export { DatePicker } from './DatePicker';
+export * from './date-picker.types';
